+++ package.json
@@ -128,6 +128,7 @@
     "wix-style-processor": "^3.1.14",
     "wix-ui-framework": "^3.0.0",
     "wix-ui-test-utils": "^1.0.0",
+    "yargs": "^16.2.0",
     "yoshi": "^4.9.0"
   },
   "husky": {
